+++ image-api/src/main/resources/no/ndla/imageapi/db/migration/V12__AddImageFileTable.sql
