+++ image-api/src/main/resources/no/ndla/imageapi/db/migration/V14__DeleteExtraImageFileData.sql
