+++ image-api/src/main/resources/no/ndla/imageapi/db/migration/V1__Create_Initial_Tables.sql
