+++ image-api/src/main/scala/no/ndla/imageapi/DBMigrator.scala
@@ -8,7 +8,7 @@
 
 package no.ndla.imageapi
 
-import imageapi.db.migrationwithdependencies.{V6__AddAgreementToImages, V7__TranslateUntranslatedAuthors}
+import no.ndla.imageapi.db.migrationwithdependencies.{V6__AddAgreementToImages, V7__TranslateUntranslatedAuthors}
 import no.ndla.imageapi.integration.DataSource
 import org.flywaydb.core.Flyway
 import org.flywaydb.core.api.output.MigrateResult
@@ -26,7 +26,7 @@ trait DBMigrator {
           new V6__AddAgreementToImages(props),
           new V7__TranslateUntranslatedAuthors(props)
         )
-        .locations("imageapi/db/migration")
+        .locations("no/ndla/imageapi/db/migration")
         .dataSource(dataSource)
         // Seems like flyway uses datasource.getConnection().getScheme() which is null if the scheme does not exist.
         // Therefore we simply override it with dataSource.getScheme.
