+++ image-api/src/main/scala/no/ndla/imageapi/db/migration/V15__OneImageFileRowForEachLanguage.scala
@@ -5,7 +5,7 @@
  * See LICENSE
  */
 
-package imageapi.db.migration
+package no.ndla.imageapi.db.migration
 
 import no.ndla.language.Language
 import org.flywaydb.core.api.migration.{BaseJavaMigration, Context}
