+++ image-api/src/main/scala/no/ndla/imageapi/db/migration/V2__RemoveFullFromImagePath.scala
@@ -6,7 +6,7 @@
  *
  */
 
-package imageapi.db.migration
+package no.ndla.imageapi.db.migration
 
 import com.typesafe.scalalogging.LazyLogging
 import org.flywaydb.core.api.migration.{BaseJavaMigration, Context}
