+++ image-api/src/main/scala/no/ndla/imageapi/db/migrationwithdependencies/V6__AddAgreementToImages.scala
@@ -5,11 +5,19 @@
  * See LICENSE
  */
 
-package imageapi.db.migrationwithdependencies
+package no.ndla.imageapi.db.migrationwithdependencies
 
 import com.typesafe.scalalogging.LazyLogging
-import imageapi.db.migration._
 import no.ndla.imageapi.ImageApiProperties
+import no.ndla.imageapi.db.migration.{
+  V5_Author,
+  V5_ImageAltText,
+  V5_ImageCaption,
+  V5_ImageMetaInformation,
+  V5_ImageTag,
+  V5_ImageTitle,
+  V5_License
+}
 import org.flywaydb.core.api.migration.{BaseJavaMigration, Context}
 import org.json4s.ext.JavaTimeSerializers
 import org.json4s.native.Serialization.{read, write}
