+++ image-api/src/test/scala/no/ndla/imageapi/db/migration/V5__AddLanguageToAllTest.scala
@@ -5,16 +5,9 @@
  * See LICENSE
  */
 
-package db.migration
+package no.ndla.imageapi.db.migration
 
-import imageapi.db.migration.{
-  V5_ImageAltText,
-  V5_ImageCaption,
-  V5_ImageMetaInformation,
-  V5_ImageTag,
-  V5_ImageTitle,
-  V5__AddLanguageToAll
-}
+import no.ndla.imageapi.db.migration.{V5_ImageAltText, V5_ImageCaption, V5_ImageMetaInformation, V5_ImageTag, V5_ImageTitle, V5__AddLanguageToAll}
 import no.ndla.imageapi.{TestEnvironment, UnitSuite}
 
 class V5__AddLanguageToAllTest extends UnitSuite with TestEnvironment {
