+++ image-api/src/test/scala/no/ndla/imageapi/db/migration/V8__UpdateLicensesTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package db.migration
+package no.ndla.imageapi.db.migration
 
-import imageapi.db.migration.V8__UpdateLicenses
+import no.ndla.imageapi.db.migration.V8__UpdateLicenses
 import no.ndla.imageapi.{TestEnvironment, UnitSuite}
 
 class V8__UpdateLicensesTest extends UnitSuite with TestEnvironment {
